+++ CHANGES.md
@@ -3,6 +3,9 @@
   * Fixed bigarray library dependencies.  No need for the user to explicitly
     link with `bigarray` when using certain functions anymore.
 
+  * Used untagged integer representations in external calls for improved
+    efficiency.
+
 ### 4.1.0 (2017-08-02)
 
   * Switched to jbuilder and topkg